+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 ### Fixed
 
 - Fixed stacking issue in Software with multiple screenshots.
-
+- Added Count label to In-Progress item.
 
 ## [0.2.8]
 ### Fixed